перенос слушателей внутрь карточки0

+++ src/components/api.js
@@ -32,12 +32,17 @@ export default class Api {
     return fetch(`${this._url}/cards/${id}`, {method: 'DELETE', headers: this._headers}).then(res => this._getResponse(res))
   }
 
-  likeCardById(id) {
-    return fetch(`${this._url}/cards/likes/${id}`, {method: 'PUT', headers: this._headers}).then(res => this._getResponse(res))
+  toggleLikeCardById(id, isLiked) {
+    if (isLiked) {
+      return fetch(`${this._url}/cards/likes/${id}`, {method: 'DELETE', headers: this._headers}).then(res => this._getResponse(res))
+    } else {
+      return fetch(`${this._url}/cards/likes/${id}`, {method: 'PUT', headers: this._headers}).then(res => this._getResponse(res))
+    }
   }
-  delLikeCardById(id) {
-    return fetch(`${this._url}/cards/likes/${id}`, {method: 'DELETE', headers: this._headers}).then(res => this._getResponse(res))
-  }
+  
+  // delLikeCardById(id) {
+  //   return fetch(`${this._url}/cards/likes/${id}`, {method: 'DELETE', headers: this._headers}).then(res => this._getResponse(res))
+  // }
   patchAvatar(userAvatar) {
     return fetch(`${this._url}/users/me/avatar`, {method: 'PATCH', headers: this._headers,
       body: JSON.stringify({avatar: userAvatar})
